backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
ebooklib==0.18
python-docx==1.1.0
reportlab==4.0.7
requests==2.31.0
